Split page parent to component
