+++ framerail/src/routes/[slug]/[...extra]/index.ts
@@ -1,2 +1,3 @@
 export { default as LayoutPane } from "./LayoutPane.svelte"
 export { default as MovePane } from "./MovePane.svelte"
+export { default as ParentPane } from "./ParentPane.svelte"
